refactor: data class는 data package로 이동. 주석추가. access modifier에 따른 메서드 위치 이동.
