+++ src/main/java/kht2199/lotto/exception/LottoListNotEmptyException.java
@@ -1,7 +1,9 @@
 package kht2199.lotto.exception;
 
+import kht2199.lotto.data.LottoList;
+
 /**
- * {@link kht2199.lotto.LottoList}를 초기화할 때, 비어있지 않을 경우 발생.
+ * {@link LottoList}를 초기화할 때, 비어있지 않을 경우 발생.
  *
  * @author heetaek.kim
  */
