+++ src/test/java/kht2199/lotto/LottoWinningResultTest.java
@@ -10,6 +10,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import kht2199.lotto.data.Lotto;
+
 class LottoWinningResultTest {
 
 	@ParameterizedTest
